Alerting User

+++ components/FavoriteComponent.js
@@ -1,12 +1,11 @@
 import React, { Component } from 'react';
-import { FlatList, View, Text } from 'react-native';
+import { FlatList, View, Text, Alert } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import { connect } from 'react-redux';
 import { Loading } from './LoadingComponent';
 import { baseUrl } from '../shared/baseUrl';
 import {deleteFavorite} from '../redux/ActionCreators';
 import Swipeout from 'react-native-swipeout';
-import { color } from 'react-native-reanimated';
 
 const mapStateToProps = state => {
     return {
@@ -29,7 +28,24 @@ class Favorites extends Component {
             const rightDelete = [{
                 text: 'Delete',
                 type: 'Delete',
-                onPress: () => this.props.deleteFavorite(item.id),
+                onPress: () => {
+                    Alert.alert(
+                        'Delete Favorite?',
+                        'Are you sure you wish to delete the favorite dish ' + item.name + '?',
+                        [
+                            { 
+                                text: 'Cancel', 
+                                onPress: () => console.log(item.name + 'Not Deleted'),
+                                style: ' cancel'
+                            },
+                            {
+                                text: 'OK',
+                                onPress: () => this.props.deleteFavorite(item.id)
+                            }
+                        ],
+                        { cancelable: false }
+                    );
+                },
                 backgroundColor: 'red'
             }];
